requirements.txt
PyBluez==0.22
PyYAML==3.13
